requirements.txt
"""
requirements.txt - 周报 Agent 依赖包
"""
anthropic==0.35.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
